v1.3.18 ------- Fixed really poorly done netcode stuff. This should fix the kartstuff desyncing, maybe collision stuff too? Fixed Megashroom's growth animation lasting too long.

+++ src/d_clisrv.c
@@ -518,9 +518,9 @@ static inline void resynch_write_player(resynch_pak *rsp, const size_t i)
 	for (j = 0; j < NUMPOWERS; ++j)
 		rsp->powers[j] = (UINT16)SHORT(players[i].powers[j]);
 	for (j = 0; j < NUMKARTSTUFF; ++j)
-		rsp->kartstuff[j] = (UINT16)SHORT(players[i].kartstuff[j]); // SRB2kart
+		rsp->kartstuff[j] = (INT32)SHORT(players[i].kartstuff[j]); // SRB2kart
 	for (j = 0; j < MAXPLAYERS; ++j)
-		rsp->collide[j] = (UINT16)SHORT(players[i].collide[j]); // SRB2kart
+		rsp->collide[j] = (UINT8)SHORT(players[i].collide[j]); // SRB2kart
 
 	// Score is resynched in the rspfirm resync packet
 	rsp->health = 0; // resynched with mo health
@@ -652,9 +652,9 @@ static void resynch_read_player(resynch_pak *rsp)
 	for (j = 0; j < NUMPOWERS; ++j)
 		players[i].powers[j] = (UINT16)SHORT(rsp->powers[j]);
 	for (j = 0; j < NUMKARTSTUFF; ++j)
-		players[i].kartstuff[j] = (UINT16)SHORT(rsp->kartstuff[j]); // SRB2kart
+		players[i].kartstuff[j] = (INT32)SHORT(rsp->kartstuff[j]); // SRB2kart
 	for (j = 0; j < MAXPLAYERS; ++j)
-		players[i].collide[j] = (UINT16)SHORT(rsp->collide[j]); // SRB2kart
+		players[i].collide[j] = (UINT8)SHORT(rsp->collide[j]); // SRB2kart
 
 	// Score is resynched in the rspfirm resync packet
 	players[i].health = rsp->health;
